temp fix for shader.js/2D.js
there's still a bug related to switching materials/modes on 3D geometry, this is just a bandaid. it's related to the fact that previously shader.js and 2D.js both used boxes, one with a shader, the other with a texture. now shader.js uses a plane instead, and no error. shame there isn't a resetShader() function like in Processing.