each sketch generates own 3D renderer
workaround for inability to reset lights and materials in p5 WEBGL. there might be an insane perf cost for doing  this but it's worth a shot. it does introduce some kind of annoying side effects but i think we can tool our way out of it. also the main renderer has been reverted to 2D so beginners can just stick to that and it's easy peasy.