add toggleable devmode
it's now possible to toggle p5's built in friendlyError system in the UI which can provide massive performance increases when disabled (10x in some cases per p5's docs).

in aleph-speak we're calling the friendlyError system "devmode". by default devmode is disabled to provide improved performance, but you can enable it while developing sketches to help you debug stuff more easily.

note: devmode is global with respect to displayWindows. it is possible to do it per-window but that doesn't seem helpful to me at present.